Downloading a patched System.IO.Compression.dll for NuGet
There appears to be an issue with Mono’s compression.  I believe it
is linked to this commit.

https://github.com/mono/mono/commit/77b034cbe78bfccfe4c1995152bd5562a1d69ef5

That commit changed the ‘version made by’ instead of the ‘version needed to extract.’
I believe this is the correct fix

https://gist.github.com/bholmes/09c9a4e5a5c9a2f14831a7b2b7ce89aa

The dll downloaded has that patch with the mono-4.4.0-branch-c7-baseline/5995f74
version.  I will work with the Mono team to get this resolved and we can remove this
when the fix propagates to the build bots officially.

Without this patch NuGet 3.4.4 creates a corrupt nupkg.