Fixes for VS 2017 to find 7zip
In VS 2017 there is no registry help to find the install root.
There is a MSBuild property we can use to point to the root.
(VsInstallRoot)  Only exists in 2017

We use this to find 7zip in...
$(VsInstallRoot)\Common7\IDE\Extensions\Xamarin.VisualStudio\7-Zip\7z.exe

If VsInstallRoot is not set this check should fail in the XamarinDownloadArchives
task and we assume that an older version of VS exists and we can use the registry.

Also added a property to allow users to override where to look.
XamarinBuildDownloadUser7ZipPath

If this moves again at the very least users can work around the issue by
setting the property in their project while we fix it again.

At this point it can be required to run VS as administrator as 7zip will fail
extracting archives that contain links.  The admin requirement should only be
needed the first time that you build (which will download and extract the files.)
VS can be restarted in non admin mode for following builds.  I am looking for a
good fix for this.