Fix the resolve path for Cecil when using VS 2017
The reference assemblies have moved in VS2017.  We now have to use
the MSBuild property VsInstallRoot that only exists in 2017.

If that path is not null or empty and the directory exists we assume 2017
and set the resolve path to the new location.  Otherwise we assume an older
version of VS and use the original path under Program Files.

I found this to be an issue with XamarinBuildiOSResourceRestore only.  However
XamarinBuildAndroidResourceRestore and XamarinBuildAndroidAarRestore should be
affected as well.  I have made the fix for all three.