[XBD] Add -sln- flag for 7zip extract
When we switched our 7zip distro on Windows from 9.2 to 16.02
a feature to handle symbolic links was introduced.  However you
need to be running as an admin to create the links.

This is causing XBD to fail on extraction of tar files that have
links.

This commit adds a the -snl- flag that tells 7zip to ignore links.

Finally if we are using an older version of Xamarin the 7zip exe is
still 9.2 and the flag does not exist.  A return code of 7 indicates
that a command line error has occurred and we assume that it is because of the -snl- flag.  We remove the flag and try again.

From 7zip help
-snl Store symbolic links as links (WIM and TAR formats only)