[xbd] Write a marker file for merged .dll’s
When merging resources into a .dll, we want to write out a marker file (`TheFile.dll.stamp` in this case) to indicate the .dll file has already been processed and we don’t need to inspect it (open it) to check to see if it needs processing.  This should save some time in the process.

I’ve also added more FileWrites outputs for the task (we’ll add the `obj/[Config]/XbdMerge` folder as well as any `.stamp` marker files we create here so that when a build is cleaned, everything is removed.

Finally, fixed typo in OverwriteSourceAssembly property.