[XBD] Added some simple AndroidManifest.xml fixups
Currently, as of 15.1, Xamarin.Android has a really simple implementation of merging AndroidManifest.xml files from referenced .aar’s into the app’s manifest.

One merge feature we need in Google Play Services is the ability for `android:name` attributes having values starting with a `.` to have the package name from the `manifest` element’s `package` attribute merged into the `android:name` attribute value.

For example, in Google Play Services Ads we may have:
```
<manifest package="com.google.android.gms.auth.api" ... >
    <application>
      <activity android:name=".signin.internal.SignInHubActivity" ... />
      ...
```

However, the merged `activity` element should end up coming in as:
```
<activity android:name="com.google.android.gms.auth.api.signin.internal.SignInHubActivity" ... />
```

Until Xamarin.Android supports this, we are adding the ability to patch the manifest on the fly when reembedding it.  We may add other fixes in the future, but for now we will fix up the name attributes.