[Mono.Posix] Change nuspec to only provide a reference for netstandard2.0 builds.
My understanding is that the ref folder will only use the dll fro builds and not runtime.
This should restrict project types that support netstandard2.0 (other than netcore) to not be able to use this.

That is what we want as we only provide runtime libraries for certain Linux flavors and macOS.