[Xamarin.Build.Download] Change iOS default to overwrite source dll
We have seen good results with Android already using this default.
Now we have noticed in iOS build logs that the source dll is showing up
in several places after it should have been changed.  It is possible that
we are (or will be soon) seeing a mix of dlls being sent to build tools.
See this bug report...

https://bugzilla.xamarin.com/show_bug.cgi?id=56498#c4