[XBD] Extract proguard.txt files manually…
… and add them to the `@(ProguardConfiguration)` item group for Xamarin.Android’s build tasks to include in proguard executions.

History:  `proguard.txt` files were _not_ being picked up from inside of .aar files (really any `LibraryProjectZip` item).

In this patch we are going to manually spot them and extract them to the same directory as we output the assembly with the merged resources to so they can be added to `@(ProguardConfiguration)` which the proguard build task will then see.