Fixing the CI builds
Mostly errors are caused by the fact that mdtool is no longer available.  Bumping the Cake Addins to pick up the fix for the IOSSolutionBuilder object.

"Release|iPhone" is no longer valid as a Configuration.  You need to specify the Configuration and Platform separately.

Also for some projects you need to force using mcs.  I think this is because of a know Roslyn issue.
https://bugzilla.xamarin.com/show_bug.cgi?id=56246