[XBD] MergeRes: Re-use test stream to write assembly
When we probe whether a file path is writeable, the FileStream used was
closed, only to have Mono.Cecil open up a new one when .Write is
invoked.

This slight refactor ensures that the resulting stream is passed in to
an overload of Write, making the process a bit more efficient. We do
manually close the FileStream here, something that was previously
handled in Mono.Cecil, since it opened its own FileStream.