[xbd] Add `ResolveAssemblyReferences` dependency
It’s possible to invoke a build of a project via msbuild which does not hit a path causing `ResolveAssemblyReferences` target to run.

Eg:
```
msbuild /t:UpdateAndroidResources /p:DesignTimeBuild=true /p:BuildingInsideVisualStudio=true /v:diag AndroidApp.csproj
```

This adds the `ResolveAssemblyReferences` target as a dependency of both the iOS and Android _master_ targets in Xamarin.Build.Download to ensure they are run and our references are resolved.