[Gigya] Fixing crashes when accessing GSAccount from accountDidLogin
When accountDidLogin is raised from Objective C the instance of the account
argument is not a GSAccount.  It is a GSResponse.  GSAccount extends GSResponse extends
GSObject.  GSObject keeps all properties of its instance stored in a NSDictionary.
GSAccount exposes helper properties around the dictionary in the base class.
This is why Objective C can treat this instance as an Account.

In managed the type of the instance is important.  This fix changes the
protocol definition to the actual parameter type on accountDidLogin and provides
a helper extension on GSObject to create a real GSAccount object from the
GSResponse that is passed in.