[Gigya] Fixing the sample
Adding Entitlements.plist to the simulator configurations.  There is a bug
requires an empty Entitlements.plist file to be a part of the app to allow
keychain calls to succeed.  This commit satisfies that requirement.

Failing keychain calls inside the Gigya SDK causes logins to fail
without warning to the user or developer.