[xbd] add exclusive file lock check to cecil ops
We have a potential race condition reported by a customer where somehow (even though it’s in a try/catch block), the `asmDefinition.Write (mergedAsmPath);` call is throwing in some cases due to a file access error (specifically file in use by another process).

There are a couple of possibilities:
 1. The file is actually still open by another msbuild invocation or another msbuild task running (either from our own tasks here, or some other part of the build process).
 2. Some previous process didn’t clean up the file completely, and it’s left in an open state.

I think 1 is most likely, since this seems to happen randomly and not always (and subsequent builds sometimes have the issue and sometimes do not).

This change opens and waits for an exclusive file lock on the assembly name + “.dll” before it will do any reading or writing of the assembly.  This should ensure that our own task is no the culprit for the race condition.