[xbd] Multiple file locking / caching strategies
This introduces caching around the use of `AssemblyName.GetAssemblyName()` calls so that we should never have to invoke that code more than once per assembly in a build.

We also use an exclusive file lock around the loading of the .dll assembly since there are other build tasks in XBD which open this assembly and lock the same file to not conflict.

Finally, we introduce a build engine caching to act as a lock around the processing of resources for an assembly so that multiple invocations for the same assembly will not try and do the same work over.