[xbd] emit filewrites always in proguard step
The proguard config inclusion task was caching proguard config files and a .stamp file to the intermediate output directory.

However, when the .stamp file was checked on subsequent builds, it was not adding the previously cached proguard config files to the filewrites upon invocation, so it seems the build process was not aware of these, causing the dex files to be recompiled more often than necessary: https://github.com/xamarin/xamarin-android/issues/1256

This changes the build so that the .stamp file contents list the proguard config files that were previously cached, so they can be appended to the filewrites when the task runs again detecting the existing stamp file.