Temporarily remove AnimationEngineListener
tl;dr: Google did something funny with AnimationEngine.AnimationEngineListener and we're going to exclude it from the binding at least for now.

Basically it seems like they used $ in the actual java source code name of the AnimationEngine$AnimationEngineListener interface and our generator can't cope with that (this is a very unorthodox thing to do in java).