[Guava] Use .targets files for the jars
Because the .dlls are not used at all, MSBuild thinks that it can discard it - taking the embedded .jars with it. Here we add a .targets file to always include the .jars, and this means that MSBuild can discard the .dlls as well as not have to extract the .jars from the .dlls - which is extra processing.