Update XBD ItemGroup name for AndroidAarLibrary
Some older nuget packages use the old itemgroup name but may also embed the .aar in the .dll.  This is now causing duplicate .jar issues at compilation for some users.

This fix is a breaking change, but in practice the old item group is not really used, the only case we might find is if someone in the wild is using this functionality.  The workaround in that case is to revert to an older version of XBD until the library using the old item group migrates to the new one.