Change guava.jar to ReferenceJar + AndroidJavaLibrary
Partially reverts: 362ed6a0e748bdd3288614b018bd62bdd7abab56

After thinking about the scenario a bit more, I realized that the
current way the Xamarin.Google.Guava NuGet package works is atypical.
Typically, when one NuGet package depends on another NuGet package, the
consuming NuGet package does not repackage the contents of the
dependency to be part of the new NuGet package.  Instead, the final
consuming application is expected to reference both NuGet packages,
either directly or transitively.

With the current approach, where the Xamarin.Google.Guava NuGet package
uses the `@(EmbeddedReferenceJar)` item type, the final consuming
application does *not* need to reference the Xamarin.Google.Guava NuGet
package at all (even transitively).  This breaks the usual conventions
for how NuGet packages behave.  That might cause some confusion for
bindings library authors.  It also makes it awkward to provide the
correct ProGuard configuration file to the final application
automatically.

Instead of including `guava.jar` as a `@(EmbeddedReferenceJar)`, revert
back to including it as a `@(ReferenceJar)`, but this time, *also*
include `guava.jar` as an `@(AndroidJavaLibrary)`.

When a bindings project references the NuGet package, the
`@(ReferenceJar)` item type will tell the build process to reference
`guava.jar`, allowing Java libraries that depend on it to be bound.  In
this scenario, the `@(AndroidJavaLibrary)` item type will be ignored.

When an application project references the NuGet package, the
`@(AndroidJavaLibrary)` item type will instead tell the build process to
package `guava.jar` as part of the application.  In this scenario, the
`@(ReferenceJar)` item type will be ignored.

Other changes:

Replace the `@(ProguardConfiguration)` item with a target that sets up
the configuration file by appending an argument to
`$(AndroidR8ExtraArguments)` instead.  This is required for now because
when R8 is configured for multidex only with no code shrinking, the
`@(ProguardConfiguration)` items are not used.