[Xamarin.Build.Download] add appropriate .NET 6 checks (#1000)
Right now Xamarin.Build.Download does not work appropriately in .NET 6
projects. It relies on the value of `$(TargetFrameworkIdentifier)`,
which will be set to `.NETCoreApp` in .NET 6.

To fix this, we can also check if `$(TargetPlatformIdentifer)` is
equal to `ios` or `android`. This should enable Xamarin.Build.Download
to work in .NET 6 projects going forward.

I also tried to fix different casing for `or` and `and` -- to always
use lowercase.